requirements.txt
Flask==2.1.2
python-dotenv==0.20.0
requests==2.27.1
